Simplify stack navigation for observations flow
